requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.88
botocore==1.42.88
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.102.0
deepdiff==9.0.0
dill==0.4.1
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
fastapi==0.135.3
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gevent==26.4.0
geventhttpclient==2.3.9
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
invoke==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
json5==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mock==4.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.24.4
openai==2.31.0
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orderly-set==5.5.0
packaging==24.2
pandas==2.3.3
paramiko==4.0.0
pathlib==1.0.1
pathos==0.3.5
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
pox==0.3.7
ppft==1.7.8
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2==2.9.11
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyNaCl==1.6.2
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
python-rapidjson==1.23
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==14.3.4
rpds-py==0.30.0
s3transfer==0.16.0
sagemaker==2.257.1
sagemaker-core==1.0.77
sagemaker-mlflow==0.2.0
sagemaker-schema-inference-artifacts==0.0.5
schema==0.7.8
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
skops==0.13.0
smdebug-rulesconfig==1.0.1
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
sympy==1.14.0
tblib==3.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
tritonclient==2.67.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.3
